Add missing values for Config runtime to avoid NPE 1. Add missing value for Config runtime, mainly the instance_distribution string. 2. Move the method of packing to instance distribution string from TopologyUtils (static method) to PackingPlan(instance method). 3. Update unit tests.
