+++ heron/scheduler-core/src/java/com/twitter/heron/scheduler/LaunchRunner.java
@@ -22,6 +22,7 @@ import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.proto.system.ExecutionEnvironment;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.Context;
+import com.twitter.heron.spi.common.Keys;
 import com.twitter.heron.spi.common.PackingPlan;
 import com.twitter.heron.spi.packing.IPacking;
 import com.twitter.heron.spi.scheduler.ILauncher;
@@ -119,8 +120,14 @@ public class LaunchRunner implements Callable<Boolean> {
     packing.initialize(config, runtime);
     PackingPlan packedPlan = packing.pack();
 
+    // Add the instanceDistribution to the runtime
+    Config ytruntime = Config.newBuilder()
+        .putAll(runtime)
+        .put(Keys.instanceDistribution(), packedPlan.getInstanceDistribution())
+        .build();
+
     // initialize the launcher
-    launcher.initialize(config, runtime);
+    launcher.initialize(config, ytruntime);
 
     Boolean result;
 
@@ -153,4 +160,5 @@ public class LaunchRunner implements Callable<Boolean> {
 
     return true;
   }
+
 }
