+++ heron/spi/src/java/com/twitter/heron/spi/utils/TopologyUtils.java
@@ -30,7 +30,6 @@ import java.util.logging.Logger;
 import com.twitter.heron.api.Config;
 import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.spi.common.Constants;
-import com.twitter.heron.spi.common.PackingPlan;
 
 /**
  * Utility to process TopologyAPI.Topology proto
@@ -197,31 +196,6 @@ public final class TopologyUtils {
       }
     }
     return ramMap;
-  }
-
-  // TODO(nbhagat): Formalize how to pass distribution to heron-executor.
-  public static String packingToString(PackingPlan packing) {
-    StringBuilder[] containerBuilder = new StringBuilder[packing.containers.size()];
-    for (PackingPlan.ContainerPlan container : packing.containers.values()) {
-      int index = Integer.parseInt(container.id);
-      containerBuilder[index - 1] = new StringBuilder();
-
-      for (PackingPlan.InstancePlan instance : container.instances.values()) {
-        String[] tokens = instance.id.split(":");
-        containerBuilder[index - 1].append(
-            String.format("%s:%s:%s:", tokens[1], tokens[2], tokens[3]));
-      }
-      containerBuilder[index - 1].deleteCharAt(containerBuilder[index - 1].length() - 1);
-    }
-
-    StringBuilder packingBuilder = new StringBuilder();
-    for (int i = 0; i < containerBuilder.length; ++i) {
-      StringBuilder builder = containerBuilder[i];
-      packingBuilder.append(String.format("%d:%s,", i + 1, builder.toString()));
-    }
-    packingBuilder.deleteCharAt(packingBuilder.length() - 1);
-
-    return packingBuilder.toString();
   }
 
   public static String formatRamMap(Map<String, Long> ramMap) {
